+++ Source/Scene/LabelCollection.js
@@ -1,5 +1,6 @@
 /*global define*/
 define([
+        '../Core/defaultValue',
         '../Core/DeveloperError',
         '../Core/defined',
         '../Core/destroyObject',
@@ -12,6 +13,7 @@ define([
         './HorizontalOrigin',
         './VerticalOrigin'
     ], function(
+        defaultValue,
         DeveloperError,
         defined,
         destroyObject,
@@ -271,6 +273,9 @@ define([
      * @alias LabelCollection
      * @constructor
      *
+     * @param {Matrix4} [options.modelMatrix=Matrix4.IDENTITY] The 4x4 transformation matrix that transforms each label from model to world coordinates.
+     * @param {Boolean} [options.debugShowBoundingVolume=false] For debugging only. Determines if this primitive's commands' bounding spheres are shown.
+     *
      * @performance For best performance, prefer a few collections, each with many labels, to
      * many collections with only a few labels each.  Avoid having collections where some
      * labels change every frame and others do not; instead, create one or more collections
@@ -295,7 +300,9 @@ define([
      *
      * @demo <a href="http://cesium.agi.com/Cesium/Apps/Sandcastle/index.html?src=Labels.html">Cesium Sandcastle Labels Demo</a>
      */
-    var LabelCollection = function() {
+    var LabelCollection = function(options) {
+        options = defaultValue(options, defaultValue.EMPTY_OBJECT);
+
         this._textureAtlas = undefined;
 
         this._billboardCollection = new BillboardCollection();
@@ -340,7 +347,19 @@ define([
          *   text     : 'Up'
          * });
          */
-        this.modelMatrix = Matrix4.clone(Matrix4.IDENTITY);
+        this.modelMatrix = Matrix4.clone(defaultValue(options.modelMatrix, Matrix4.IDENTITY));
+
+        /**
+         * This property is for debugging only; it is not for production use nor is it optimized.
+         * <p>
+         * Draws the bounding sphere for each {@see DrawCommand} in the primitive.
+         * </p>
+         *
+         * @type {Boolean}
+         *
+         * @default false
+         */
+        this.debugShowBoundingVolume = defaultValue(options.debugShowBoundingVolume, false);
     };
 
     /**
@@ -556,6 +575,7 @@ define([
         var billboardCollection = this._billboardCollection;
 
         billboardCollection.modelMatrix = this.modelMatrix;
+        billboardCollection.debugShowBoundingVolume = this.debugShowBoundingVolume;
 
         if (!defined(this._textureAtlas)) {
             this._textureAtlas = context.createTextureAtlas();
@@ -586,7 +606,7 @@ define([
         }
         labelsToUpdate.length = 0;
 
-        this._billboardCollection.update(context, frameState, commandList);
+        billboardCollection.update(context, frameState, commandList);
     };
 
     /**